refactor: move all vehicle component into a vehicle folder
