+++ frontend/src/app/page.tsx
@@ -1,4 +1,4 @@
-import VehiculeList from "@/components/VehicleList";
+import VehiculeList from "@/components/vehicle/VehicleList";
 
 export default async function Home() {
