+++ frontend/src/components/vehicle/VehicleCard.tsx
